backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
aiohttp>=3.8.1
networkx>=2.6.3
python-dotenv>=0.19.0
pydantic>=1.8.2
numpy>=1.24.0
requests>=2.26.0
asyncio>=3.4.3